Added a rushed trend calculation that we can fix, added urls and sources, convert percents to percents and round them to 2 decimals.

+++ MHTCDashboard/src/main/java/edu/wpi/mhtc/model/Data/Data.java
@@ -3,15 +3,15 @@ package edu.wpi.mhtc.model.Data;
 public class Data
 {
 	private final Integer year; //after 2000... so 2014 is 14, 1999 is -1;
-	private final Integer value;
+	private final double value;
 	
-	public Data(int year, int value)
+	public Data(int year, double value)
 	{
 		this.year = year;
 		this.value = value;
 	}
 
-	public Integer getValue() {
+	public double getValue() {
 		return value;
 	}
